requirements.txt
chromadb==0.4.22
datasets==2.7.1
gdown==4.6.0
ipyannotations==0.5.1
keybert==0.7.0
matplotlib==3.7.2
numpy==1.24.3
pandas==1.5.3
Pillow==10.2.0
requests==2.31.0
seaborn==0.13.2
spacy
tqdm==4.65.0
transformers==4.37.1